[mac] associate .plottr with Plottr
refs #28
on macOS, the .plottr files will have a cool Plottr icon and you can
open Plottr by double clicking them

+++ bin/build.js
@@ -15,10 +15,12 @@ var options = {
   version: '1.3.2',
   icon: 'icons/pblack',
   out: path.resolve('/', 'Users', 'csutter', 'plottr_dist'),
-  protocol: ['plottr'],
   'app-copyright': 'Copyright 2016 C. Louis S. (Cameron Sutter)',
   'app-category-type': 'public.app-category.productivity',
+  protocol: ['plottr'],
   'protocol-name': ['Plottr'],
+  'extend-info': 'extend.plist',
+  'extra-resource': 'icons/plottrfile.icns',
   arch: 'x64',
   platform: 'darwin',
   ignore: [
@@ -34,20 +36,20 @@ var options = {
 packager(options, function (errMac, appPaths) {
   if (errMac) {
     console.log('error building macOS:' + errMac, errMac.stack)
-    fs.unlinkSync('trialmode.json')
+    if (process.argv[2] === 'TRIALMODE') fs.unlinkSync('trialmode.json')
   } else {
     // win64
     options.platform = 'win32'
     packager(options, function (errWin64, appPaths) {
       if (errWin64) {
         console.log('error building win64:' + errWin64)
-        fs.unlinkSync('trialmode.json')
+        if (process.argv[2] === 'TRIALMODE') fs.unlinkSync('trialmode.json')
       } else {
         // win32
         options.arch = 'ia32'
         packager(options, function (errWin32, appPaths) {
           if (errWin32) console.log('error building win32:' + errWin32)
-          fs.unlinkSync('trialmode.json')
+          if (process.argv[2] === 'TRIALMODE') fs.unlinkSync('trialmode.json')
         })
       }
     })
